pyproject.toml
[tool.poetry]
name = "SDLon"
version = "0.0.0"
description = "OS2mo SDLoen integration"
authors = ["Magenta <info@magenta.dk>"]
license = "MPL-2.0"
homepage = "https://magenta.dk/"
repository = "https://git.magenta.dk/rammearkitektur/os2mo-data-import-and-export"
keywords = ["os2mo", "dipex", "sdloen"]

[tool.poetry.dependencies]
another-sd-client = "^1.9.0"
python = "^3.11"
more-itertools = "^8.12.0"
requests = "^2.27.1"
xmltodict = "^0.12.0"
ra-utils = "^1.7.0"
pydantic = "^1.9.0"
anytree = "^2.8.0"
fastapi = "^0.75.0"
lxml = "^4.8.0"
os2mo-data-import = "^3.33.2"
ramodels = "^5.4.0"
os2mo-dar-client = "^1.0.0"
tqdm = "^4.63.0"
uvicorn = "^0.17.6"
pytest = "6.2.5"
dipex = "^3.1.1"
click = "^8.1.0"
prometheus-fastapi-instrumentator = "^5.7.1"
prometheus-client = "^0.14.1"
sentry-sdk = "^1.12.1"
gql = "^3.4.0"
os2mo-http-trigger-protocol = "0.0.3"
os2mo-fastapi-utils = "0.0.3"
sqlalchemy = "<=2.0.0"
alembic = "<=2.0.0"

[tool.poetry.group.test.dependencies]
pytest = "^6.2.5"
hypothesis = "^6.39.3"
reorder-python-imports = "^3.0.1"
parameterized = "^0.7.4"
pre-commit = "^3.2.2"
pytest-cov = "^4.1.0"


[tool.poetry.group.pre-commit.dependencies]
mypy = "1.5.1"
black = "^22.1.0"

[tool.flake8]
max-line-length = 88

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "strict"
